server

+++ src/modules/watchlist/watchlist.service.ts
@@ -1,30 +1,41 @@
 import { Injectable } from '@nestjs/common';
+import { InjectModel } from "@nestjs/sequelize";
+import { CreateAssetResponse } from "./response";
 import { Watchlist } from './models/watchlist.module';
-import { InjectModel } from '@nestjs/sequelize';
-import { CreateAssetResponse } from './response';
-import { WatchListDTO } from './dto';
 
 @Injectable()
 export class WatchlistService {
-  constructor(
-    @InjectModel(Watchlist)
-    private readonly watchlistRepository: typeof Watchlist,
-  ) {}
+  constructor(@InjectModel(Watchlist) private readonly watchlistRepository: typeof Watchlist) {}
 
-  async createAsset (user: { id: any; }, dto: WatchListDTO): Promise<CreateAssetResponse> {
-    const watchlist = {
-      user: user.id,
-      name: dto.name,
-      assetId: dto.assetId
+  async createAsset (user, dto): Promise<CreateAssetResponse> {
+    try {
+      const watchlist = {
+        user: user.id,
+        name: dto.name,
+        assetId: dto.assetId
+      }
+      await this.watchlistRepository.create(watchlist)
+      
+      return watchlist
+    }catch (e) {
+      throw new Error(e)
     }
-    await this.watchlistRepository.create(watchlist)
-    return watchlist
   }
 
-  async deleteAsset(userId: number, assetId: string): Promise<boolean> {
-    await this.watchlistRepository.destroy({
-      where: { id: assetId, user: userId },
-    });
-    return true;
+  async getUserAssets (userId: number): Promise<Watchlist[]> {
+    try {
+      return this.watchlistRepository.findAll({where: {user: userId}})
+    }catch (e) {
+      throw new Error(e)
+    }
+  }
+
+  async deleteAsset (userId: number, assetId: string): Promise<boolean> {
+    try {
+      await this.watchlistRepository.destroy({where: {id: assetId, user: userId}})
+      return true
+    }catch (e) {
+      throw new Error(e)
+    }
   }
 }
